Imported long-stack-traces lib to figure out crashes.
Timeouts are happening. It seems if I keep the server up
and running for more yhan a day or so, it timesout. Since
node handles events asynchronously, it becomes difficult to
trace the stack trace to see where in the code is faulty
because JavaScript stack traces are limited to a single event.
Long stacktraces allows me to capture such traces and hopefully
figure out where the crash happens.